packages/sr2/pyproject.toml
[project]
name = "sr2"
version = "1.0.0"
description = "Context engineering library for AI agents"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
a2a = ["a2a-sdk>=0.3"]
postgres = ["asyncpg>=0.29"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.scripts]
sr2-config-docs = "schema_gen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sr2", "src/schema_gen.py"]
